Update waste-detection docs/tests and bump trainer rev
